Implementando arquivo conftest.py
