+++ libpythonpro/tests/test_spam/test_envio_para_base_de_usuarios.py
@@ -1,0 +1,11 @@
+from libpythonpro.spam.enviador_de_email import Enviador
+from libpythonpro.spam.main import EnviadorDeSpam
+
+
+def test_envio_de_spam(sessao):
+    enviador_de_spam= EnviadorDeSpam(sessao, Enviador())
+    enviador_de_spam.enviar_emails(
+        'atiladalan@gmail.com',
+        'Curso Python Pro',
+        'Confira os Modulos fantasticos'
+    )
